More tests fixed. These are tests that cannot work in RegTest, only in MainNet, because of the very low minimum difficulty of RegTest.
