+++ rskj-core/src/test/java/co/rsk/mine/MinerServerTest.java
@@ -187,63 +187,6 @@ public class MinerServerTest {
     }
 
 
-    /*
-     * This test is much more likely to fail than the
-     * submitBitcoinBlockProofOfWorkNotGoodEnough test. Even then
-     * it should almost never fail.
-     */
-    @Test
-    public void submitBitcoinBlockInvalidBlockDoesntEliminateCache() {
-        /* We need a low, but not too low, target */
-        BlockChainImpl bc = new BlockChainBuilder().build();
-        Genesis gen = (Genesis) BlockChainImplTest.getGenesisBlock(bc);
-        gen.getHeader().setDifficulty(BigInteger.valueOf(300000).toByteArray());
-        bc.setStatus(gen, gen.getCumulativeDifficulty());
-        World world = new World(bc, gen);
-        blockchain = world.getBlockChain();
-
-        EthereumImpl ethereumImpl = Mockito.mock(EthereumImpl.class);
-        Mockito.when(ethereumImpl.addNewMinedBlock(Mockito.any())).thenReturn(ImportResult.IMPORTED_BEST);
-
-        BlockUnclesValidationRule unclesValidationRule = Mockito.mock(BlockUnclesValidationRule.class);
-        Mockito.when(unclesValidationRule.isValid(Mockito.any())).thenReturn(true);
-        MinerServer minerServer = new MinerServerImpl(ethereumImpl, this.blockchain, null,
-                this.blockchain.getPendingState(), blockchain.getRepository(),
-                ConfigUtils.getDefaultMiningConfig(), unclesValidationRule,
-                world.getBlockProcessor(), DIFFICULTY_CALCULATOR,
-                new GasLimitCalculator(RskSystemProperties.CONFIG),
-                new ProofOfWorkRule(RskSystemProperties.CONFIG).setFallbackMiningEnabled(false));
-
-        minerServer.start();
-        MinerWork work = minerServer.getWork();
-
-        co.rsk.bitcoinj.core.BtcBlock bitcoinMergedMiningBlock = getMergedMiningBlock(work);
-
-        bitcoinMergedMiningBlock.setNonce(1);
-
-        // Try to submit a block with invalid PoW, this should not eliminate the block from the cache
-        SubmitBlockResult result1 = minerServer.submitBitcoinBlock(work.getBlockHashForMergedMining(), bitcoinMergedMiningBlock);
-
-        Assert.assertEquals("ERROR", result1.getStatus());
-        Assert.assertNull(result1.getBlockInfo());
-        Mockito.verify(ethereumImpl, Mockito.times(0)).addNewMinedBlock(Mockito.any());
-
-        // Now try to submit the same block, this should work fine since the block remains in the cache
-        findNonce(work, bitcoinMergedMiningBlock);
-
-        SubmitBlockResult result2 = minerServer.submitBitcoinBlock(work.getBlockHashForMergedMining(), bitcoinMergedMiningBlock);
-
-        Assert.assertEquals("OK", result2.getStatus());
-        Assert.assertNotNull(result2.getBlockInfo());
-        Mockito.verify(ethereumImpl, Mockito.times(1)).addNewMinedBlock(Mockito.any());
-
-        // Finally, submit the same block again and validate that addNewMinedBlock is called again
-        SubmitBlockResult result3 = minerServer.submitBitcoinBlock(work.getBlockHashForMergedMining(), bitcoinMergedMiningBlock);
-
-        Assert.assertEquals("OK", result3.getStatus());
-        Assert.assertNotNull(result3.getBlockInfo());
-        Mockito.verify(ethereumImpl, Mockito.times(2)).addNewMinedBlock(Mockito.any());
-    }
 
     @Test
     public void workWithNoTransactionsZeroFees() {
